Create sampling_and_time_summarization
Lines 50 through 54 will sample the data.  I nested this with Jon's original cleaning data to run this independently of other code.  56 through 82 is code that will provide you a summary of the data for that sample.  Comment out the sample and add in your timeframe to do period analysis.  You can adjust the sample size on line 52.